Removed the typescript declaration from package.json. Added background color to the alert box. Refactored AlertBox to it's own component. Added invisible div so user can click anywhere on the screen to remove the alert box.
