+++ package.json
@@ -17,6 +17,7 @@
     "react": "^16.12.0",
     "react-beautiful-dnd": "^12.2.0",
     "react-dom": "^16.12.0",
+    "react-onclickoutside": "^6.9.0",
     "react-router-dom": "^5.1.2",
     "react-scripts": "^3.4.3",
     "typeface-roboto": "0.0.75",
